test: rename _fixtures to __fixtures__
