+++ .prettierignore
@@ -10,6 +10,5 @@
 /*.log
 package.json
 **/__fixtures__/
-**/_fixtures/
 **/__mocks__/
 bin/yarn*
