+++ test/config/config/__fixtures__/file.js
