+++ test/config/config/__fixtures__/file2.js
