+++ test/config/npm/__fixtures__/coffeelint.json
