+++ test/config/npm/__fixtures__/next.json
