+++ test/config/npm/__fixtures__/vue.json
