+++ test/config/npm/__fixtures__/webpack.json
