+++ test/platform/bitbucket-server/__fixtures__/responses.ts
