+++ test/platform/bitbucket/__fixtures__/responses.ts
