+++ test/platform/bitbucket/comments.spec.ts
@@ -1,7 +1,7 @@
 import URL from 'url';
 import { api as _api } from '../../../lib/platform/bitbucket/bb-got-wrapper';
 import * as comments from '../../../lib/platform/bitbucket/comments';
-import responses from './_fixtures/responses';
+import responses from './__fixtures__/responses';
 
 jest.mock('../../../lib/platform/bitbucket/bb-got-wrapper');
