+++ test/platform/github/__fixtures__/graphql/pullrequests-closed.json
