+++ test/platform/github/index.spec.ts
@@ -69,11 +69,11 @@ describe('platform/github', () => {
   });
 
   const graphqlOpenPullRequests = fs.readFileSync(
-    'test/platform/github/_fixtures/graphql/pullrequest-1.json',
+    'test/platform/github/__fixtures__/graphql/pullrequest-1.json',
     'utf8'
   );
   const graphqlClosedPullrequests = fs.readFileSync(
-    'test/platform/github/_fixtures/graphql/pullrequests-closed.json',
+    'test/platform/github/__fixtures__/graphql/pullrequests-closed.json',
     'utf8'
   );
