+++ test/platform/utils/__fixtures__/issue-body.txt
