+++ test/workers/repository/process/lookup/index.spec.ts
@@ -1,12 +1,12 @@
 import nock from 'nock';
 import * as lookup from '../../../../../lib/workers/repository/process/lookup';
-import qJson from '../../../../config/npm/_fixtures/01.json';
-import helmetJson from '../../../../config/npm/_fixtures/02.json';
-import coffeelintJson from '../../../../config/npm/_fixtures/coffeelint.json';
-import webpackJson from '../../../../config/npm/_fixtures/webpack.json';
-import nextJson from '../../../../config/npm/_fixtures/next.json';
-import vueJson from '../../../../config/npm/_fixtures/vue.json';
-import typescriptJson from '../../../../config/npm/_fixtures/typescript.json';
+import qJson from '../../../../config/npm/__fixtures__/01.json';
+import helmetJson from '../../../../config/npm/__fixtures__/02.json';
+import coffeelintJson from '../../../../config/npm/__fixtures__/coffeelint.json';
+import webpackJson from '../../../../config/npm/__fixtures__/webpack.json';
+import nextJson from '../../../../config/npm/__fixtures__/next.json';
+import vueJson from '../../../../config/npm/__fixtures__/vue.json';
+import typescriptJson from '../../../../config/npm/__fixtures__/typescript.json';
 import * as _docker from '../../../../../lib/datasource/docker';
 import * as _gitSubmodules from '../../../../../lib/datasource/git-submodules';
 import { mocked, getConfig } from '../../../../util';
